UserController and Token changes

+++ Dasa-Challenge-BackEnd/src/controller/userController.js
@@ -1,7 +1,9 @@
 const User = require("../models/userModel");
 const PasswordToken = require("../models/passwordTokenModel");
+const { v4: uuidv4 } = require("uuid");
 const bcrypt = require("bcrypt");
 const jwt = require("jsonwebtoken");
+const { Sequelize } = require("sequelize");
 const secret = "adsuasgdhjasgdhjdgahjsg12hj3eg12hj3g12hj3g12hj3g123";
 
 class UserController {
@@ -36,6 +38,11 @@ class UserController {
       const { user_email, user_name, user_password } = req.body;
       let { user_image } = req.body;
 
+      if (!user_password || user_password.trim() === "") {
+        res.status(400).json({ error: "A senha é obrigatória!" });
+        return;
+      }
+
       if (!user_email || user_email.trim() === "") {
         res.status(400).json({ error: "O e-mail é inválido!" });
         return;
@@ -63,8 +70,17 @@ class UserController {
 
       res.status(200).json({ message: "Usuário criado com sucesso!" });
     } catch (error) {
-      console.error(error);
-      res.status(500).json({ error: "Erro ao criar usuário." });
+      if (error.name === "SequelizeValidationError") {
+        // Mapeia os erros para uma resposta JSON
+        const errors = error.errors.map((err) => ({
+          field: err.path,
+          message: err.message,
+        }));
+        res.status(400).json({ errors });
+      } else {
+        console.error(error);
+        res.status(500).json({ error: "Erro ao criar usuário." });
+      }
     }
   }
 
@@ -87,7 +103,13 @@ class UserController {
 
   async remove(req, res) {
     const id = req.params.user_id;
+    const requestingUserId = req.user.user_id;
+    const requestingUserRole = req.user.user_role;
 
+    console.log(typeof requestingUserId,typeof parseInt(id), requestingUserRole)
+    if(requestingUserId !== parseInt(id) && requestingUserRole !== 1){
+      return res.status(403).json({error: "Você não tem permissão para deletar esse usuário!"})
+    }
     try {
       const result = await User.destroy({ where: { user_id: id } });
       if (result) {
@@ -104,33 +126,49 @@ class UserController {
   async recoverPassword(req, res) {
     const { user_email } = req.body;
     try {
-      const result = await PasswordToken.create(user_email);
-      if (result.status) {
-        res.status(200).send("" + result.token);
-      } else {
-        res.status(406).send(result.error);
+      const token = uuidv4();
+      const user = await User.findOne({ where: { user_email: user_email } });
+
+      if (!user) {
+        return res.status(404).send("Usuário não encontrado");
       }
+      const result = await PasswordToken.create(
+        {
+          user_id: user.user_id,
+          token: token,
+          used: false,
+        },
+        { expiresIn: "1h" }
+      );
+
+      res.status(200).json({token: result.token});
     } catch (error) {
-      res.status(500).send("Ocorreu um erro no servidor.");
+      if (error instanceof Sequelize.ValidationError) {
+        return res.status(400).send(error.errors);
+      }
+      res.status(500).send(error);
     }
   }
 
   async changePassword(req, res) {
     const { token, user_password } = req.body;
+
     try {
       const isTokenValid = await PasswordToken.validate(token);
+
       if (isTokenValid.status) {
         const hashedPassword = await bcrypt.hash(user_password, 10);
         await User.update(
           { user_password: hashedPassword },
-          { where: { token_id: isTokenValid.token.user_id } }
+          { where: { user_id: isTokenValid.token.user_id } }
         );
+        await PasswordToken.markAsUsed(token);
         res.status(200).send("Senha alterada com sucesso!");
       } else {
         res.status(406).send("Token inválido!");
       }
     } catch (error) {
-      res.status(500).send("Ocorreu um erro no servidor.");
+      res.status(500).send({ error });
     }
   }
